refactor: 프로필 등록 page 컴포넌트 분리
